Custom sliders, fix issues, and make compilable.
Add my custom class KnobLookAndFeel, inherited from JUCE's LookAndFeel.
Use this class and a custom image strip to create a custom slider/knob.

Fix some issues regarding identifiers matching across files and headers
not being included. Fix everything to make the IntroJucer generate the
correct, needed code. Make the project able to compile and test if the
window and knobs and displaying/functioning correctly.

Next step:
Probably create the callback for the sliders and do the maintenance
needed for using the sliders with the AudioProcessor.