Create the Transposed Direct Form II Filter
Create the class for the Transposed Direct Form II Filter and create the
variables in the AudioProcessor class
(TransposedDirectFormIifilterAudioProcessor) for the filter (type,
cutoff, Q, etc.).

Note:
Filter will still need to be implemented in the processBlock() function.
Also, the GUI controls will need to be configured in the AudioProcessor
class, "synced" with the filter's variables, and updated with the
timerCallback().