Register UI/parameter elements in PluginProcessor
Register UI/parameter elements in PluginProcessor. Write things like
methods for storing parameters in memory (using XML). Also write code
for parameters in methods like getParameter(), setParameter(), etc. This
gets it ready to talk to the AudioProcessorEditor so the UI and
processor can "communicate" with each other. This also allows to save
and recall "states" of the plugin in memory.

Note:
I still need to "connect" and update the UI in the AudioProcessorEditor.
For instance, I need to write the necessary code in timerCallback() to
periodically check if any of the parameters have changed.