Implement most of the filter, except for filter type control
Implement almost everything the filter needs. Add the update methods for
"communicating" between the Editor and AudioProcessor, which will
periodically check if parameters have changed in the Editor. Make other
small additions where needed.

Note:
The UI control for changing Filter Type needs to be added to the Editor.
The filter won't currently work without it because it doesn't know what
type of filter is suppose to be used. It just send through unprocessed
signal.