Use proper cutoff and use Res instead of Q
Add equations for calculating the proper Cutoff frequency and
calculating Q from a Resonance parameter (range: 0-1). Change the names
and values of the sliders to correspond to these changes.

Also, change the names in the TransDirectFormIIFilter class (methods,
arguments, etc.) to fit with the new Resonance parameter. Do the same
thing in the PluginProcessor and wherever else needed. Comment the
member functions in the TransDirectFormIIFilter class to very briefly
describe their purpose.

Give the filter the ability to instantiate itself when the plugin is
started so that it will actually process audio without clicking on a
knob or anything. (Before, a knob had to be clicked on or turned for the
audio to be processed). This is accomplished in either the
AudioProcessor's constructor or in AudioProcessor::prepareToPlay().

Note:
Next Goal - Make the filter knob change the cutoff frequency
logarithmically, converting pitch values (0-127) to frequency values
(Hz). This makes for much more natural control, especially for lower
frequencies because it has more precision the lower it is.