Document filter and make other small improvements
Document (comment) the filter's member functions to where a description
will show up in Intellisense.

Create a file that will be used for useful DSP tasks, called
DspUtilities. Use the pitchToFreq() function made (DspUtilities) in
TransDirectFormIIFilter to convert from pitch (MIDI note #) to
frequency.

Change the cutoffSlider's range from frequency values to pitch (MIDI
note #) values. Using pitch values gives the control a much more natural
feel when sweeping and has more accuracy at lower values.

Create tooltips for the sliders to give a very brief description of the
parameters they're changing. Also, add proper suffixes to the labels of
the corresponding slider ("Hz", "dB").

Add a file called Parameters that will be used to handle the parameters
from the plugin. The way the code is currently written uses old methods
that will eventually be deprecated.

Note:
The class for the parameters still needs to be written and the current
code to handle parameters from the plugin and to/from the host needs to
be changed to use the newer way of handling parameters. Also, the
parameters need to be optimized to send values from 0-1 so it can
correctly "communicate" with the host.