Make 1 function for processing, rename, etc.
Create a function that can be used for stereo processing (2 channels),
instead of using two separate functions for left and right. This is done
by using arrays for the filter's states (z^-1).

Rename some of the member functions and arguments to more closely
represent what they are.

Update to JUCE 3.1.1